+++ src/Oro/Bundle/ApiBundle/Tests/Unit/ApiDoc/Parser/ApiDocMetadataParserTest.php
@@ -38,7 +38,7 @@ class ApiDocMetadataParserTest extends \PHPUnit_Framework_TestCase
     public function testSupportsWithoutMetadata()
     {
         $item = [
-            'class'   => 'Test\Class',
+            'class'   => null,
             'options' => []
         ];
 
@@ -48,7 +48,7 @@ class ApiDocMetadataParserTest extends \PHPUnit_Framework_TestCase
     public function testSupportsWithMetadata()
     {
         $item = [
-            'class'   => ApiDocMetadata::class,
+            'class'   => null,
             'options' => [
                 'metadata' => new ApiDocMetadata(
                     'test',
@@ -65,26 +65,9 @@ class ApiDocMetadataParserTest extends \PHPUnit_Framework_TestCase
     public function testSupportsWithUnknownMetadata()
     {
         $item = [
-            'class'   => 'Test\Class',
+            'class'   => null,
             'options' => [
                 'metadata' => new \stdClass()
-            ]
-        ];
-
-        $this->assertFalse($this->parser->supports($item));
-    }
-
-    public function testSupportsWithMetadataButWithInvalidClassName()
-    {
-        $item = [
-            'class'   => 'Test\Class',
-            'options' => [
-                'metadata' => new ApiDocMetadata(
-                    'test',
-                    $this->createMock(EntityMetadata::class),
-                    $this->createMock(EntityDefinitionConfig::class),
-                    new RequestType([])
-                )
             ]
         ];