+++ src/Oro/Bundle/ApiBundle/Tests/Unit/Processor/OptimizedProcessorBagTest.php
@@ -5,6 +5,7 @@ namespace Oro\Bundle\ApiBundle\Tests\Unit\Processor;
 use Oro\Component\ChainProcessor\ChainApplicableChecker;
 use Oro\Component\ChainProcessor\Context;
 use Oro\Component\ChainProcessor\ProcessorApplicableCheckerFactoryInterface;
+use Oro\Component\ChainProcessor\ProcessorBagConfigBuilder;
 use Oro\Component\ChainProcessor\ProcessorFactoryInterface;
 use Oro\Component\ChainProcessor\ProcessorIterator;
 use Oro\Component\ChainProcessor\ProcessorIteratorFactoryInterface;
@@ -53,7 +54,9 @@ class OptimizedProcessorBagTest extends \PHPUnit_Framework_TestCase
             ->method('createApplicableChecker')
             ->willReturn($this->ungroupedApplicableChecker);
 
+        $processorBagConfigBuilder = new ProcessorBagConfigBuilder();
         $this->processorBag = new OptimizedProcessorBag(
+            $processorBagConfigBuilder,
             $this->processorFactory,
             false,
             $this->applicableCheckerFactory,
@@ -62,9 +65,9 @@ class OptimizedProcessorBagTest extends \PHPUnit_Framework_TestCase
             $this->ungroupedProcessorIteratorFactory
         );
 
-        $this->processorBag->addGroup('group1', 'action_with_groups');
-        $this->processorBag->addProcessor('processor1', [], 'action_with_groups', 'group1');
-        $this->processorBag->addProcessor('processor2', [], 'action_without_groups');
+        $processorBagConfigBuilder->addGroup('group1', 'action_with_groups');
+        $processorBagConfigBuilder->addProcessor('processor1', [], 'action_with_groups', 'group1');
+        $processorBagConfigBuilder->addProcessor('processor2', [], 'action_without_groups');
     }
 
     public function testBagWithGroups()
@@ -77,7 +80,7 @@ class OptimizedProcessorBagTest extends \PHPUnit_Framework_TestCase
         $this->processorIteratorFactory->expects(self::once())
             ->method('createProcessorIterator')
             ->with(
-                [['processor' => 'processor1', 'attributes' => ['group' => 'group1']]],
+                [['processor1', ['group' => 'group1']]],
                 self::identicalTo($context),
                 self::identicalTo($this->applicableChecker),
                 self::identicalTo($this->processorFactory)
@@ -100,7 +103,7 @@ class OptimizedProcessorBagTest extends \PHPUnit_Framework_TestCase
         $this->ungroupedProcessorIteratorFactory->expects(self::once())
             ->method('createProcessorIterator')
             ->with(
-                [['processor' => 'processor2', 'attributes' => []]],
+                [['processor2', []]],
                 self::identicalTo($context),
                 self::identicalTo($this->ungroupedApplicableChecker),
                 self::identicalTo($this->processorFactory)