+++ src/Oro/Component/ChainProcessor/ProcessorBagInterface.php
@@ -2,28 +2,11 @@
 
 namespace Oro\Component\ChainProcessor;
 
+/**
+ * Represents a read-only collection of processors.
+ */
 interface ProcessorBagInterface
 {
-    /**
-     * Registers a processing group.
-     *
-     * @param string $group
-     * @param string $action
-     * @param int    $priority
-     */
-    public function addGroup($group, $action, $priority = 0);
-
-    /**
-     * Registers a processor.
-     *
-     * @param string      $processorId
-     * @param array       $attributes
-     * @param string|null $action
-     * @param string|null $group
-     * @param int         $priority
-     */
-    public function addProcessor($processorId, array $attributes, $action = null, $group = null, $priority = 0);
-
     /**
      * Registers a processor applicable checker.
      *