+++ src/Oro/Component/ChainProcessor/ProcessorIterator.php
@@ -7,16 +7,7 @@ namespace Oro\Component\ChainProcessor;
  */
 class ProcessorIterator implements \Iterator
 {
-    /**
-     * @var array
-     *  [
-     *      [
-     *          'processor'  => processorId,
-     *          'attributes' => [key => value, ...]
-     *      ],
-     *      ...
-     *  ]
-     */
+    /** @var array [[processor id, [attribute name => attribute value, ...]], ...] */
     protected $processors;
 
     /** @var ContextInterface */
@@ -93,7 +84,7 @@ class ProcessorIterator implements \Iterator
             return null;
         }
 
-        $attributes = $this->processors[$this->index]['attributes'];
+        $attributes = $this->processors[$this->index][1];
 
         if (!isset($attributes['group'])) {
             return null;
@@ -113,7 +104,7 @@ class ProcessorIterator implements \Iterator
             return null;
         }
 
-        return $this->processors[$this->index]['processor'];
+        return $this->processors[$this->index][0];
     }
 
     /**
@@ -127,7 +118,7 @@ class ProcessorIterator implements \Iterator
             return null;
         }
 
-        return $this->processors[$this->index]['attributes'];
+        return $this->processors[$this->index][1];
     }
 
     /**
@@ -135,7 +126,7 @@ class ProcessorIterator implements \Iterator
      */
     public function current()
     {
-        $processorId = $this->processors[$this->index]['processor'];
+        $processorId = $this->processors[$this->index][0];
         $processor = $this->processorFactory->getProcessor($processorId);
         if (null === $processor) {
             throw new \RuntimeException(sprintf('The processor "%s" does not exist.', $processorId));
@@ -187,7 +178,7 @@ class ProcessorIterator implements \Iterator
         while ($this->index <= $this->maxIndex) {
             $applicable = $this->applicableChecker->isApplicable(
                 $this->context,
-                $this->processors[$this->index]['attributes']
+                $this->processors[$this->index][1]
             );
             if ($applicable !== ApplicableCheckerInterface::NOT_APPLICABLE) {
                 break;