+++ src/Oro/Component/ChainProcessor/Tests/Unit/GroupRangeApplicableCheckerTest.php
@@ -4,6 +4,7 @@ namespace Oro\Component\ChainProcessor\Tests\Unit;
 
 use Oro\Component\ChainProcessor\Context;
 use Oro\Component\ChainProcessor\ProcessorBag;
+use Oro\Component\ChainProcessor\ProcessorBagConfigBuilder;
 use Oro\Component\ChainProcessor\ProcessorFactoryInterface;
 
 class GroupRangeApplicableCheckerTest extends \PHPUnit_Framework_TestCase
@@ -13,17 +14,18 @@ class GroupRangeApplicableCheckerTest extends \PHPUnit_Framework_TestCase
         $context = new Context();
         $context->setAction('action1');
 
-        $processorBag = new ProcessorBag($this->getProcessorFactory());
+        $builder = new ProcessorBagConfigBuilder();
+        $processorBag = new ProcessorBag($builder, $this->getProcessorFactory());
 
-        $processorBag->addGroup('group1', 'action1', -10);
+        $builder->addGroup('group1', 'action1', -10);
 
-        $processorBag->addProcessor('processor1_no_action', []);
-        $processorBag->addProcessor('processor2_no_action', [], null, null, -1);
-        $processorBag->addProcessor('processor1_no_group', [], 'action1');
-        $processorBag->addProcessor('processor2_no_group', [], 'action1', null, -1);
-        $processorBag->addProcessor('processor1', [], 'action1', 'group1');
+        $builder->addProcessor('processor1_no_action', []);
+        $builder->addProcessor('processor2_no_action', [], null, null, -1);
+        $builder->addProcessor('processor1_no_group', [], 'action1');
+        $builder->addProcessor('processor2_no_group', [], 'action1', null, -1);
+        $builder->addProcessor('processor1', [], 'action1', 'group1');
 
-        $this->assertProcessors(
+        self::assertProcessors(
             [
                 'processor1_no_action',
                 'processor1_no_group',
@@ -41,17 +43,18 @@ class GroupRangeApplicableCheckerTest extends \PHPUnit_Framework_TestCase
         $context->setAction('action1');
         $context->setFirstGroup('unknown_group');
 
-        $processorBag = new ProcessorBag($this->getProcessorFactory());
+        $builder = new ProcessorBagConfigBuilder();
+        $processorBag = new ProcessorBag($builder, $this->getProcessorFactory());
 
-        $processorBag->addGroup('group1', 'action1', -10);
+        $builder->addGroup('group1', 'action1', -10);
 
-        $processorBag->addProcessor('processor1_no_action', []);
-        $processorBag->addProcessor('processor2_no_action', [], null, null, -1);
-        $processorBag->addProcessor('processor1_no_group', [], 'action1');
-        $processorBag->addProcessor('processor2_no_group', [], 'action1', null, -1);
-        $processorBag->addProcessor('processor1', [], 'action1', 'group1');
+        $builder->addProcessor('processor1_no_action', []);
+        $builder->addProcessor('processor2_no_action', [], null, null, -1);
+        $builder->addProcessor('processor1_no_group', [], 'action1');
+        $builder->addProcessor('processor2_no_group', [], 'action1', null, -1);
+        $builder->addProcessor('processor1', [], 'action1', 'group1');
 
-        $this->assertProcessors(
+        self::assertProcessors(
             [
                 'processor1_no_action',
                 'processor1_no_group',
@@ -69,17 +72,18 @@ class GroupRangeApplicableCheckerTest extends \PHPUnit_Framework_TestCase
         $context->setAction('action1');
         $context->setLastGroup('unknown_group');
 
-        $processorBag = new ProcessorBag($this->getProcessorFactory());
+        $builder = new ProcessorBagConfigBuilder();
+        $processorBag = new ProcessorBag($builder, $this->getProcessorFactory());
 
-        $processorBag->addGroup('group1', 'action1', -10);
+        $builder->addGroup('group1', 'action1', -10);
 
-        $processorBag->addProcessor('processor1_no_action', []);
-        $processorBag->addProcessor('processor2_no_action', [], null, null, -1);
-        $processorBag->addProcessor('processor1_no_group', [], 'action1');
-        $processorBag->addProcessor('processor2_no_group', [], 'action1', null, -1);
-        $processorBag->addProcessor('processor1', [], 'action1', 'group1');
+        $builder->addProcessor('processor1_no_action', []);
+        $builder->addProcessor('processor2_no_action', [], null, null, -1);
+        $builder->addProcessor('processor1_no_group', [], 'action1');
+        $builder->addProcessor('processor2_no_group', [], 'action1', null, -1);
+        $builder->addProcessor('processor1', [], 'action1', 'group1');
 
-        $this->assertProcessors(
+        self::assertProcessors(
             [
                 'processor1_no_action',
                 'processor1_no_group',
@@ -98,27 +102,28 @@ class GroupRangeApplicableCheckerTest extends \PHPUnit_Framework_TestCase
         $context->setFirstGroup('group2');
         $context->setLastGroup('group5');
 
-        $processorBag = new ProcessorBag($this->getProcessorFactory());
+        $builder = new ProcessorBagConfigBuilder();
+        $processorBag = new ProcessorBag($builder, $this->getProcessorFactory());
 
-        $processorBag->addGroup('group1', 'action1', -10);
-        $processorBag->addGroup('group2', 'action1', -20);
-        $processorBag->addGroup('group4', 'action1', -30);
-        $processorBag->addGroup('group3', 'action1', -40);
-        $processorBag->addGroup('group5', 'action1', -50);
-        $processorBag->addGroup('group6', 'action1', -60);
+        $builder->addGroup('group1', 'action1', -10);
+        $builder->addGroup('group2', 'action1', -20);
+        $builder->addGroup('group4', 'action1', -30);
+        $builder->addGroup('group3', 'action1', -40);
+        $builder->addGroup('group5', 'action1', -50);
+        $builder->addGroup('group6', 'action1', -60);
 
-        $processorBag->addProcessor('processor1_no_action', []);
-        $processorBag->addProcessor('processor2_no_action', [], null, null, -1);
-        $processorBag->addProcessor('processor1_no_group', [], 'action1');
-        $processorBag->addProcessor('processor2_no_group', [], 'action1', null, -1);
-        $processorBag->addProcessor('processor1', [], 'action1', 'group1');
-        $processorBag->addProcessor('processor2', [], 'action1', 'group2');
-        $processorBag->addProcessor('processor3', [], 'action1', 'group3');
-        $processorBag->addProcessor('processor4', [], 'action1', 'group4');
-        $processorBag->addProcessor('processor5', [], 'action1', 'group5');
-        $processorBag->addProcessor('processor6', [], 'action1', 'group6');
+        $builder->addProcessor('processor1_no_action', []);
+        $builder->addProcessor('processor2_no_action', [], null, null, -1);
+        $builder->addProcessor('processor1_no_group', [], 'action1');
+        $builder->addProcessor('processor2_no_group', [], 'action1', null, -1);
+        $builder->addProcessor('processor1', [], 'action1', 'group1');
+        $builder->addProcessor('processor2', [], 'action1', 'group2');
+        $builder->addProcessor('processor3', [], 'action1', 'group3');
+        $builder->addProcessor('processor4', [], 'action1', 'group4');
+        $builder->addProcessor('processor5', [], 'action1', 'group5');
+        $builder->addProcessor('processor6', [], 'action1', 'group6');
 
-        $this->assertProcessors(
+        self::assertProcessors(
             [
                 'processor1_no_action',
                 'processor1_no_group',
@@ -139,27 +144,28 @@ class GroupRangeApplicableCheckerTest extends \PHPUnit_Framework_TestCase
         $context->setAction('action1');
         $context->setFirstGroup('group2');
 
-        $processorBag = new ProcessorBag($this->getProcessorFactory());
+        $builder = new ProcessorBagConfigBuilder();
+        $processorBag = new ProcessorBag($builder, $this->getProcessorFactory());
 
-        $processorBag->addGroup('group1', 'action1', -10);
-        $processorBag->addGroup('group2', 'action1', -20);
-        $processorBag->addGroup('group4', 'action1', -30);
-        $processorBag->addGroup('group3', 'action1', -40);
-        $processorBag->addGroup('group5', 'action1', -50);
-        $processorBag->addGroup('group6', 'action1', -60);
+        $builder->addGroup('group1', 'action1', -10);
+        $builder->addGroup('group2', 'action1', -20);
+        $builder->addGroup('group4', 'action1', -30);
+        $builder->addGroup('group3', 'action1', -40);
+        $builder->addGroup('group5', 'action1', -50);
+        $builder->addGroup('group6', 'action1', -60);
 
-        $processorBag->addProcessor('processor1_no_action', []);
-        $processorBag->addProcessor('processor2_no_action', [], null, null, -1);
-        $processorBag->addProcessor('processor1_no_group', [], 'action1');
-        $processorBag->addProcessor('processor2_no_group', [], 'action1', null, -1);
-        $processorBag->addProcessor('processor1', [], 'action1', 'group1');
-        $processorBag->addProcessor('processor2', [], 'action1', 'group2');
-        $processorBag->addProcessor('processor3', [], 'action1', 'group3');
-        $processorBag->addProcessor('processor4', [], 'action1', 'group4');
-        $processorBag->addProcessor('processor5', [], 'action1', 'group5');
-        $processorBag->addProcessor('processor6', [], 'action1', 'group6');
+        $builder->addProcessor('processor1_no_action', []);
+        $builder->addProcessor('processor2_no_action', [], null, null, -1);
+        $builder->addProcessor('processor1_no_group', [], 'action1');
+        $builder->addProcessor('processor2_no_group', [], 'action1', null, -1);
+        $builder->addProcessor('processor1', [], 'action1', 'group1');
+        $builder->addProcessor('processor2', [], 'action1', 'group2');
+        $builder->addProcessor('processor3', [], 'action1', 'group3');
+        $builder->addProcessor('processor4', [], 'action1', 'group4');
+        $builder->addProcessor('processor5', [], 'action1', 'group5');
+        $builder->addProcessor('processor6', [], 'action1', 'group6');
 
-        $this->assertProcessors(
+        self::assertProcessors(
             [
                 'processor1_no_action',
                 'processor1_no_group',
@@ -181,27 +187,28 @@ class GroupRangeApplicableCheckerTest extends \PHPUnit_Framework_TestCase
         $context->setAction('action1');
         $context->setLastGroup('group5');
 
-        $processorBag = new ProcessorBag($this->getProcessorFactory());
+        $builder = new ProcessorBagConfigBuilder();
+        $processorBag = new ProcessorBag($builder, $this->getProcessorFactory());
 
-        $processorBag->addGroup('group1', 'action1', -10);
-        $processorBag->addGroup('group2', 'action1', -20);
-        $processorBag->addGroup('group4', 'action1', -30);
-        $processorBag->addGroup('group3', 'action1', -40);
-        $processorBag->addGroup('group5', 'action1', -50);
-        $processorBag->addGroup('group6', 'action1', -60);
+        $builder->addGroup('group1', 'action1', -10);
+        $builder->addGroup('group2', 'action1', -20);
+        $builder->addGroup('group4', 'action1', -30);
+        $builder->addGroup('group3', 'action1', -40);
+        $builder->addGroup('group5', 'action1', -50);
+        $builder->addGroup('group6', 'action1', -60);
 
-        $processorBag->addProcessor('processor1_no_action', []);
-        $processorBag->addProcessor('processor2_no_action', [], null, null, -1);
-        $processorBag->addProcessor('processor1_no_group', [], 'action1');
-        $processorBag->addProcessor('processor2_no_group', [], 'action1', null, -1);
-        $processorBag->addProcessor('processor1', [], 'action1', 'group1');
-        $processorBag->addProcessor('processor2', [], 'action1', 'group2');
-        $processorBag->addProcessor('processor3', [], 'action1', 'group3');
-        $processorBag->addProcessor('processor4', [], 'action1', 'group4');
-        $processorBag->addProcessor('processor5', [], 'action1', 'group5');
-        $processorBag->addProcessor('processor6', [], 'action1', 'group6');
+        $builder->addProcessor('processor1_no_action', []);
+        $builder->addProcessor('processor2_no_action', [], null, null, -1);
+        $builder->addProcessor('processor1_no_group', [], 'action1');
+        $builder->addProcessor('processor2_no_group', [], 'action1', null, -1);
+        $builder->addProcessor('processor1', [], 'action1', 'group1');
+        $builder->addProcessor('processor2', [], 'action1', 'group2');
+        $builder->addProcessor('processor3', [], 'action1', 'group3');
+        $builder->addProcessor('processor4', [], 'action1', 'group4');
+        $builder->addProcessor('processor5', [], 'action1', 'group5');
+        $builder->addProcessor('processor6', [], 'action1', 'group6');
 
-        $this->assertProcessors(
+        self::assertProcessors(
             [
                 'processor1_no_action',
                 'processor1_no_group',
@@ -238,7 +245,7 @@ class GroupRangeApplicableCheckerTest extends \PHPUnit_Framework_TestCase
      * @param string[]  $expectedProcessorIds
      * @param \Iterator $processors
      */
-    protected function assertProcessors(array $expectedProcessorIds, \Iterator $processors)
+    protected static function assertProcessors(array $expectedProcessorIds, \Iterator $processors)
     {
         $processorIds = [];
         /** @var ProcessorMock $processor */
@@ -246,6 +253,6 @@ class GroupRangeApplicableCheckerTest extends \PHPUnit_Framework_TestCase
             $processorIds[] = $processor->getProcessorId();
         }
 
-        $this->assertEquals($expectedProcessorIds, $processorIds);
+        self::assertEquals($expectedProcessorIds, $processorIds);
     }
 }